# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VentureLog API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JournalTag = (typeof JournalTag)[keyof typeof JournalTag];

export const JournalTag = {
  Product: "Product",
  Growth: "Growth",
  Team: "Team",
  Fundraising: "Fundraising",
  Operations: "Operations",
  Finance: "Finance",
  Reflection: "Reflection",
} as const;

export interface JournalEntry {
  id: string;
  content: string;
  tags: JournalTag[];
  isPromoted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateJournalEntryBody {
  content: string;
  tags: JournalTag[];
  isPromoted?: boolean;
}

export interface UpdateJournalEntryBody {
  content?: string;
  tags?: JournalTag[];
  isPromoted?: boolean;
}

export type MetricClass = (typeof MetricClass)[keyof typeof MetricClass];

export const MetricClass = {
  revenue: "revenue",
  retention: "retention",
  engagement: "engagement",
  unit_economics: "unit_economics",
} as const;

export type MetricPeriod = (typeof MetricPeriod)[keyof typeof MetricPeriod];

export const MetricPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;

export type MetricDirection =
  (typeof MetricDirection)[keyof typeof MetricDirection];

export const MetricDirection = {
  higher_is_better: "higher_is_better",
  lower_is_better: "lower_is_better",
  context_dependent: "context_dependent",
} as const;

export interface Metric {
  id: string;
  name: string;
  class: MetricClass;
  period: MetricPeriod;
  direction: MetricDirection;
  formulaNumerator?: string | null;
  formulaDenominator?: string | null;
  isArchived: boolean;
  createdAt: string;
  currentValue?: number | null;
  previousValue?: number | null;
  lastUpdatedAt?: string | null;
}

export interface MetricValue {
  id: string;
  metricId: string;
  value: number;
  recordedDate: string;
  isCorrection: boolean;
  originalValue?: number | null;
  correctionNote?: string | null;
  createdAt: string;
}

export type DecisionStatus =
  (typeof DecisionStatus)[keyof typeof DecisionStatus];

export const DecisionStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface DecisionLogItem {
  id: string;
  sourceEntryId?: string | null;
  linkedMetricId?: string | null;
  linkedMetricName?: string | null;
  title: string;
  contextSummary: string;
  optionsConsidered: string[];
  chosenOption: string;
  expectedOutcome?: string | null;
  actualOutcome?: string | null;
  lessonsLearned?: string | null;
  tags: JournalTag[];
  status: DecisionStatus;
  hasComments: boolean;
  isArchived: boolean;
  createdAt: string;
}

export interface MetricDetail {
  metric: Metric;
  values: MetricValue[];
  linkedDecisions: DecisionLogItem[];
}

export interface CreateMetricBody {
  name: string;
  class: MetricClass;
  period: MetricPeriod;
  direction: MetricDirection;
  formulaNumerator?: string | null;
  formulaDenominator?: string | null;
}

export interface UpdateMetricBody {
  name?: string;
  isArchived?: boolean;
}

export interface LogMetricValueBody {
  value: number;
  recordedDate: string;
  isCorrection?: boolean;
  originalValue?: number | null;
  correctionNote?: string | null;
}

export interface DecisionComment {
  id: string;
  decisionId: string;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface DecisionDetail {
  decision: DecisionLogItem;
  comments: DecisionComment[];
}

export interface CreateDecisionBody {
  sourceEntryId?: string | null;
  linkedMetricId?: string | null;
  /** @maxLength 120 */
  title: string;
  /** @maxLength 500 */
  contextSummary: string;
  optionsConsidered: string[];
  chosenOption: string;
  expectedOutcome?: string | null;
  tags: JournalTag[];
}

export interface UpdateDecisionBody {
  /** @maxLength 120 */
  title?: string;
  /** @maxLength 500 */
  contextSummary?: string;
  optionsConsidered?: string[];
  chosenOption?: string;
  expectedOutcome?: string | null;
  actualOutcome?: string | null;
  lessonsLearned?: string | null;
  linkedMetricId?: string | null;
  tags?: JournalTag[];
  status?: DecisionStatus;
  isArchived?: boolean;
}

export interface CreateDecisionCommentBody {
  authorName: string;
  content: string;
}

export interface SparklinePoint {
  date: string;
  value: number;
}

export interface MetricSparkline {
  metricId: string;
  metricName: string;
  currentValue?: number | null;
  previousValue?: number | null;
  direction: MetricDirection;
  class: MetricClass;
  sparkline: SparklinePoint[];
}

export type DailyPromptType =
  (typeof DailyPromptType)[keyof typeof DailyPromptType];

export const DailyPromptType = {
  stale_metric: "stale_metric",
  open_decision: "open_decision",
  default: "default",
} as const;

export interface DailyPrompt {
  type: DailyPromptType;
  message: string;
  hasEntryToday: boolean;
}

export interface DashboardSummary {
  prompt: DailyPrompt;
  topMetrics: MetricSparkline[];
  recentDecisions: DecisionLogItem[];
  totalJournalEntries: number;
  totalDecisions: number;
  openDecisions: number;
}

export type ListJournalEntriesParams = {
  tag?: string;
  search?: string;
};

export type ListDecisionsParams = {
  status?: ListDecisionsStatus;
  tag?: string;
  search?: string;
  includeArchived?: boolean;
};

export type ListDecisionsStatus =
  (typeof ListDecisionsStatus)[keyof typeof ListDecisionsStatus];

export const ListDecisionsStatus = {
  open: "open",
  closed: "closed",
} as const;